+++ irisData/multinomialBayes.py
@@ -7,13 +7,15 @@ with open('iris_shuffled.csv', 'rt') as csvfile:
     for row in csvObj:
         data.append(row)
 data = np.array(data)
-trainData = data[:100, :4]
+trainData = data[:100, :4].astype(np.float)
 trainLabel = data[:100, 4]
-testData = data[100:, :4]
+testData = data[100:, :4].astype(np.float)
 testLabel = data[100:, 4]
 clf = multinomialBayes(fit_prior = 0.5)
 clf.fit(trainData, trainLabel)
 print('Score: ' + str(clf.score(testData, testLabel)))
+# print('Predict: ' + str(clf.predict(testData)))
+# print('Predict Probability: ' + str(clf.predictProbability()))
 # saveModel(clf, 'model.sav') #save model for future use
 # loaded_model = loadModel('model.sav') #load model from file
 # print(loaded_model.predict(testData)) #try to predict using loaded_model
